Add files via upload
language_analysis.ipynb is a jupyter notebook containing the code for isolating, cleaning, and basic analysis of the survey respondents answers to which programming languages they use.

"20**_languages_isolated" are files containing the raw data from the Stack Overflow User Survey.  The full files were too big to upload, so these are just the isolated columns.